+++ src/Checkout/CheckoutException.php
@@ -16,8 +16,6 @@ use Symfony\Component\HttpFoundation\Response;
 class CheckoutException extends HttpException
 {
     public const PREPARED_ORDER_REQUIRED = 'PREPARED_ORDER_REQUIRED';
-    public const EXPRESS_MISSING_COUNTRY_CODE = 'EXPRESS_MISSING_COUNTRY_CODE';
-    public const EXPRESS_COUNTRY_NOT_FOUND = 'EXPRESS_COUNTRY_NOT_FOUND';
 
     /**
      * @param class-string<AbstractPaymentHandler> $paymentHandler
@@ -29,25 +27,6 @@ class CheckoutException extends HttpException
             self::PREPARED_ORDER_REQUIRED,
             'PayPal Order ID does not exist in the request. The payment method {{ paymentHandler }} requires a prepared PayPal order.',
             ['paymentHandler' => $paymentHandler],
-        );
-    }
-
-    public static function expressMissingCountryCode(): self
-    {
-        return new self(
-            Response::HTTP_BAD_REQUEST,
-            self::EXPRESS_MISSING_COUNTRY_CODE,
-            'Missing country code in shipping address',
-        );
-    }
-
-    public static function expressCountryNotFound(string $countryCode): self
-    {
-        return new self(
-            Response::HTTP_BAD_REQUEST,
-            self::EXPRESS_COUNTRY_NOT_FOUND,
-            'Country not found for code: {{ countryCode }}',
-            ['countryCode' => $countryCode],
         );
     }
 }