+++ src/Resources/app/administration/src/types/openapi.d.ts
@@ -1294,15 +1294,21 @@ export interface components {
             full_name: string;
         };
         paypal_v1_token: {
-            /** @description Scopes expressed in the form of resource URL endpoints. The value of the scope parameter
-             *     is expressed as a list of space-delimited, case-sensitive strings. */
+            /**
+             * @description Scopes expressed in the form of resource URL endpoints. The value of the scope parameter
+             *     is expressed as a list of space-delimited, case-sensitive strings.
+             */
             scope: string;
             nonce: string;
-            /** @description The access token issued by PayPal. After the access token
-             *     expires (see $expiresIn), you must request a new access token. */
+            /**
+             * @description The access token issued by PayPal. After the access token
+             *     expires (see $expiresIn), you must request a new access token.
+             */
             access_token: string;
-            /** @description The type of the token issued as described in OAuth2.0 RFC6749,
-             *     Section 7.1. Value is case insensitive. */
+            /**
+             * @description The type of the token issued as described in OAuth2.0 RFC6749,
+             *     Section 7.1. Value is case insensitive.
+             */
             token_type: string;
             app_id: string;
             id_token: string | null;
@@ -1356,16 +1362,23 @@ export interface components {
             state: string;
             merchant_id: string | null;
         };
+        paypal_v1_webhook_list: {
+            webhooks: components["schemas"]["paypal_v1_webhook"][];
+        };
         paypal_v2_common_address: {
-            /** @description The first line of the address. For example, number or street. For example, 173 Drury Lane.
-             *     Required for data entry and compliance and risk checks. Must contain the full address. */
+            /**
+             * @description The first line of the address. For example, number or street. For example, 173 Drury Lane.
+             *     Required for data entry and compliance and risk checks. Must contain the full address.
+             */
             address_line_1: string | null;
             /** @description The second line of the address. For example, suite or apartment number. */
             address_line_2: string | null;
             /** @description A city, town, or village. Smaller than $adminArea1 */
             admin_area_2: string | null;
-            /** @description The highest level sub-division in a country, which is usually a province, state, or ISO-3166-2 subdivision.
-             *     Format for postal delivery. For example, CA and not California. */
+            /**
+             * @description The highest level sub-division in a country, which is usually a province, state, or ISO-3166-2 subdivision.
+             *     Format for postal delivery. For example, CA and not California.
+             */
             admin_area_1: string | null;
             postal_code: string | null;
             country_code: string;
@@ -1676,6 +1689,7 @@ export interface components {
             shipping: components["schemas"]["paypal_v2_order_purchase_unit_shipping"];
             payments: components["schemas"]["paypal_v2_order_purchase_unit_payments"] | null;
             supplementary_data: components["schemas"]["paypal_v2_order_purchase_unit_supplementary_data"];
+            shipping_options: components["schemas"]["paypal_v2_order_purchase_unit_shipping_option"][] | null;
         };
         paypal_v2_order_purchase_unit_amount: components["schemas"]["paypal_v2_common_money"] & {
             breakdown: components["schemas"]["paypal_v2_order_purchase_unit_amount_breakdown"] | null;
@@ -1797,6 +1811,14 @@ export interface components {
             url: string | null;
             image_url: string | null;
         };
+        paypal_v2_order_purchase_unit_shipping_option: {
+            id: string;
+            label: string;
+            amount: components["schemas"]["paypal_v2_common_money"];
+            /** @enum {string} */
+            type: "SHIPPING" | "PICKUP";
+            selected: boolean;
+        };
         paypal_v2_order_purchase_unit_supplementary_data: {
             card: components["schemas"]["paypal_v2_order_purchase_unit_supplementary_data_card"];
             risk: components["schemas"]["paypal_v2_order_purchase_unit_supplementary_data_risk"];
@@ -1845,6 +1867,12 @@ export interface components {
             /** @default false */
             notify_payer: boolean;
             items: components["schemas"]["paypal_v2_order_purchase_unit_shipping_tracker_item"][];
+        };
+        paypal_v2_order_shipping_callback: {
+            id: string;
+            shipping_address: components["schemas"]["paypal_v2_common_address"];
+            shipping_option: components["schemas"]["paypal_v2_order_purchase_unit_shipping_option"];
+            purchase_units: components["schemas"]["paypal_v2_order_purchase_unit"][];
         };
         paypal_v2_patch: {
             op: string;