+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+# 10.2.0
 - Adjusted the amount that Pay Later is available for to the current limitations given (https://developer.paypal.com/studio/checkout/pay-later/de)
 - Fixes an issue, where Paypal Express Checkout does not recalculate cart price after changing shipping country (shopware/SwagPayPal#342)
+- Fixes an issue, where the Smart Payment Buttons were not stretched to full width on the product detail page for digital products
+- Fixes an issue, where the webhook registration could run into a loop, if the saved Webhook ID is invalid
+- Fixed an issue, where Paypal errors showing on category page in Storefront, which should only be showed on checkout pages (shopware/shopware#12701)
 
 # 10.1.2
 - Fixes an issue, where the plugin could not be updated when Zettle Sales Channels were active (shopware/SwagPayPal#361)